requirements.txt
streamlit
pandas
# yfinance
scikit-learn
plotly
matplotlib
seaborn
streamlit-lottie
requests
smartapi-python
pyotp